strat_signal
