+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: junseyun <junseyun@student.42gyeongsan.    +#+  +:+       +#+         #
+#    By: dohyuki2 <dohyuki2@student.42Gyeongsan.    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/19 13:00:35 by dohyuki2          #+#    #+#              #
-#    Updated: 2024/12/26 14:34:24 by junseyun         ###   ########.fr        #
+#    Updated: 2024/12/26 20:13:50 by dohyuki2         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -42,6 +42,7 @@ SRCS = minishell.c \
 		built_in/pwd.c \
 		built_in/unset.c \
 		parse/check_param.c \
+		parse/check_param2.c \
 		parse/tokenize.c \
 		parse/substitution.c \
 		parse/substitution2.c \
