+++ Utilities/signal.c
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dohyuki2 <dohyuki2@student.42Gyeongsan.    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/12/26 20:04:53 by dohyuki2          #+#    #+#             */
+/*   Updated: 2024/12/26 20:10:27 by dohyuki2         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../minishell.h"
+
+void	signal_handler(int sig)
+{
+	if (sig == SIGINT)
+		write(1, "^C\n", 3);
+}
