+++ parse/check_param.c
@@ -6,7 +6,7 @@
 /*   By: dohyuki2 <dohyuki2@student.42Gyeongsan.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/12 16:29:17 by dohyuki2          #+#    #+#             */
-/*   Updated: 2024/12/24 15:29:00 by dohyuki2         ###   ########.fr       */
+/*   Updated: 2024/12/26 20:13:19 by dohyuki2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,6 @@
 
 int		check_quote(char *param, char c);
 int		check_grammer(char *param);
-void	jump_quote(char *param, int *i);
-void	jump_sp(char *param, int *i);
 int		check_pipe_parse(char *param, int *i);
 int		check_redirect(char *param, int *i);
 
@@ -98,33 +96,4 @@ int	check_pipe_parse(char *param, int *i)
 			return (1);
 	}
 	return (0);
-}
-
-void	jump_sp(char *param, int *i)
-{
-	while (param[*i] == 32 && param[*i] != '\0')
-		++*i;
-	return ;
-}
-void	jump_quote(char *param, int *i)
-{
-	if (param[*i] == 34)
-	{
-		*i = *i + 1;
-		while (param[*i] != 34 && param[*i] != '\0')
-			++*i;
-		if (param[*i] == '\0')
-			--*i;
-		return ;
-	}
-	if (param[*i] == 39)
-	{
-		*i = *i + 1;
-		while (param[*i] != 39 && param[*i] != '\0')
-			++*i;
-		if (param[*i] == '\0')
-			--*i;
-		return ;
-	}
-	return ;
 }
